backend/requirements/base.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
pydantic>=2.0.0
cryptography>=42.0.0
python-decouple>=3.8
django-cors-headers>=4.3.0
jsonschema>=4.20.0
httpx>=0.25.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0
fastmcp>=2.0.0
authlib>=1.3.0
pyjwt>=2.9.0
jinja2>=3.1.0
boto3>=1.34.0
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-instrumentation-fastapi>=0.45b0

